Kotlin multiplatform for common data transfer objects
