+++ build.gradle.kts
@@ -1,6 +1,7 @@
 plugins {
     kotlin("jvm") version "1.4.10" apply false
     kotlin("js") version "1.4.10" apply false
+    kotlin("multiplatform") version "1.4.10" apply false
     kotlin("plugin.spring") version "1.4.10" apply false
     kotlin("plugin.serialization") version "1.4.10" apply false
     id("org.springframework.boot") version "2.3.4.RELEASE" apply false
