+++ reporting/common/src/commonMain/kotlin/com/datasite/poc/garden/report/dto/ReportEvent.kt
@@ -1,0 +1,13 @@
+package com.datasite.poc.garden.report.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+sealed class ReportEvent {
+    @Serializable
+    @SerialName("dummy")
+    class DummyEvent(val character: Char, val count: Long) : ReportEvent()
+
+    // TODO build out report event types
+}
