+++ reporting/service/src/main/kotlin/com/datasite/poc/garden/report/entity/DummyEntity.kt
@@ -1,0 +1,18 @@
+package com.datasite.poc.garden.report.entity
+
+import com.datasite.poc.garden.report.dto.Dummy
+import org.springframework.data.annotation.Id
+import org.springframework.data.relational.core.mapping.Table
+
+const val DUMMY_TABLE = "dummies"
+
+@Table(DUMMY_TABLE)
+class DummyEntity(
+    @Id val character: Char,
+    val count: Long
+)
+
+fun DummyEntity.toDummy() = Dummy(
+    character,
+    count,
+)
