--- reporting/src/main/kotlin/com/datasite/poc/garden/reporting/DatasiteGardenReportApplication.kt
@@ -1,16 +1,0 @@
-package com.datasite.poc.garden.reporting
-
-import kotlinx.serialization.json.Json
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-import org.springframework.context.annotation.Bean
-
-@SpringBootApplication
-class DatasiteGardenReportApplication {
-    @get:Bean
-    val json: Json = Json.Default
-}
-
-fun main(args: Array<String>) {
-    runApplication<DatasiteGardenReportApplication>(*args)
-}
