+++ reporting/ui/src/main/kotlin/com/datasite/poc/garden/report/App.kt
@@ -1,0 +1,15 @@
+package com.datasite.poc.garden.report
+
+import com.bnorm.react.RFunction
+import react.RBuilder
+import react.dom.div
+
+@Suppress("FunctionName")
+@RFunction
+fun RBuilder.App() {
+    +"Hello, World!"
+
+    div {
+        DummyReportTable()
+    }
+}
