+++ reporting/ui/src/main/kotlin/com/datasite/poc/garden/report/index.kt
@@ -1,0 +1,10 @@
+package com.datasite.poc.garden.report
+
+import kotlinx.browser.document
+import react.dom.render
+
+fun main() {
+    document.getElementById("root")?.let {
+        render(it) { App() }
+    }
+}
