+++ reporting/ui/src/main/resources/index.html
