+++ settings.gradle.kts
@@ -1,6 +1,7 @@
 rootProject.name = "datasite-poc-garden"
 
 include(":processor")
-include(":reporting")
+include(":reporting:common")
+include(":reporting:service")
+include(":reporting:ui")
 include(":service")
-include(":ui")
