--- ui/src/main/kotlin/com/datasite/poc/garden/ui/App.kt
@@ -1,15 +1,0 @@
-package com.datasite.poc.garden.ui
-
-import com.bnorm.react.RFunction
-import react.RBuilder
-import react.dom.div
-
-@Suppress("FunctionName")
-@RFunction
-fun RBuilder.App() {
-    +"Hello, World!"
-
-    div {
-        DummyReportTable()
-    }
-}
