--- ui/src/main/kotlin/com/datasite/poc/garden/ui/Dummy.kt
@@ -1,16 +1,0 @@
-package com.datasite.poc.garden.ui
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Dummy(
-    val character: Char,
-    val count: Long
-) {
-    companion object {
-        fun from(event: ReportEvent.DummyEvent) = Dummy(
-            event.character,
-            event.count,
-        )
-    }
-}
