--- ui/src/main/kotlin/com/datasite/poc/garden/ui/index.kt
@@ -1,10 +1,0 @@
-package com.datasite.poc.garden.ui
-
-import kotlinx.browser.document
-import react.dom.render
-
-fun main() {
-    document.getElementById("root")?.let {
-        render(it) { App() }
-    }
-}
