--- ui/src/main/kotlin/com/datasite/poc/garden/ui/react.kt
@@ -1,19 +1,0 @@
-package com.datasite.poc.garden.ui
-
-import io.ktor.client.features.websocket.*
-import io.ktor.client.request.*
-import io.ktor.http.cio.websocket.*
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.launch
-import react.RDependenciesList
-import react.useEffectWithCleanup
-
-// React hook to run an async process during the component lifecycle
-fun useAsync(dependencies: RDependenciesList? = null, block: suspend () -> Unit) {
-    useEffectWithCleanup(dependencies) {
-        val job = GlobalScope.launch {
-            block()
-        }
-        return@useEffectWithCleanup { job.cancel() }
-    }
-}
